requirements.txt
feedparser==6.0.2
python-dotenv==0.15.0